Feat: 루틴 조회 api v2 연동 (#T3-159)
