+++ Projects/Domain/Sources/Protocol/Repository/RoutineRepositoryProtocol.swift
@@ -21,9 +21,7 @@ public protocol RoutineRepositoryProtocol {
     /// - Parameters:
     ///   - startDate: 조회 시작 날짜
     ///   - endDate: 조회 종료 날짜
-    func fetchRoutines(from startDate: String, to endDate: String) async throws -> [String: [RoutineEntity]]
-
-    func fetchRoutines2(from startDate: String, to endDate: String) async throws -> [String: (routine: [newRoutineEntity], allCompleted: Bool)]
+    func fetchRoutines(from startDate: String, to endDate: String) async throws -> [String: (routines: [RoutineEntity], allCompleted: Bool)]
 
     /// 루틴을 수정합니다.
     /// - Parameters:
