+++ Projects/Presentation/Sources/Home/ViewModel/HomeViewModel.swift
@@ -17,7 +17,7 @@ final class HomeViewModel: ViewModel {
         case selectDate(date: Date)
         case selectRoutineListDate
         case fetchRoutines
-        case selectRoutine(routine: MainRoutine?)
+        case selectRoutine(routine: Routine?)
         case deleteDailyRoutine
         case deleteAllRoutine
         case updateRoutineCompletion(updatedRoutine: Routine)
@@ -30,20 +30,20 @@ final class HomeViewModel: ViewModel {
         let selectedDatePublisher: AnyPublisher<Date, Never>
         let routineListDatePublisher: AnyPublisher<Date, Never>
         let fetchRoutineResultPublisher: AnyPublisher<Bool, Never>
-        let routinesPublisher: AnyPublisher<[MainRoutine], Never>
+        let routinesPublisher: AnyPublisher<[Routine], Never>
         let deleteRoutineResultPublisher: AnyPublisher<Bool, Never>
         let updateRoutineCompletionResultPublisher: AnyPublisher<Bool, Never>
     }
 
     private(set) var output: Output
-    private var routines: [String: [MainRoutine]] = [:]
+    private var routines: [String: [Routine]] = [:]
     private let nicknameSubject = CurrentValueSubject<String, Never>("")
     private let emotionSubject = CurrentValueSubject<Emotion?, Never>(nil)
     private let selectedDateSubject = CurrentValueSubject<Date, Never>(.now)
     private let routineListDateSubject = PassthroughSubject<Date, Never>()
     private let fetchRoutineResultSubject = PassthroughSubject<Bool, Never>()
-    private let routinesSubject = CurrentValueSubject<[MainRoutine], Never>([])
-    private let selectedRoutineSubject = CurrentValueSubject<MainRoutine?, Never>(nil)
+    private let routinesSubject = CurrentValueSubject<[Routine], Never>([])
+    private let selectedRoutineSubject = CurrentValueSubject<Routine?, Never>(nil)
     private let deleteRoutineResultSubject = PassthroughSubject<Bool, Never>()
     private let updateRoutineCompletionResultSubject = PassthroughSubject<Bool, Never>()
 
@@ -198,8 +198,9 @@ final class HomeViewModel: ViewModel {
         Task {
             do {
                 let entities = try await routineUseCase.fetchRoutines(startDate: startDate, endDate: endDate)
-                for (date, routineEntities) in entities {
-                    routines[date] = routineEntities.compactMap({ $0.toMainRoutine() })
+                for (date, values) in entities {
+                    let routineEntities = values.routines
+                    routines[date] = routineEntities.compactMap({ $0.toRoutine() })
                 }
                 fetchRoutineResultSubject.send(true)
             } catch {
@@ -227,6 +228,7 @@ final class HomeViewModel: ViewModel {
 
     // 당일 루틴을 삭제합니다.
     private func deleteDailyRoutine() {
+        /*
         guard let routine = selectedRoutineSubject.value
         else { return }
 
@@ -249,10 +251,12 @@ final class HomeViewModel: ViewModel {
                 deleteRoutineResultSubject.send(false)
             }
         }
+         */
     }
 
     // 루틴의 완료 여부를 업데이트 합니다.
     private func updateRoutineCompletion(updatedRoutine: Routine) {
+        /*
         let performedDate = selectedDateSubject.value.convertToString(dateType: .yearMonthDate)
         var routineCompletionEntities: [RoutineCompletionEntity] = []
 
@@ -311,5 +315,6 @@ final class HomeViewModel: ViewModel {
                 updateRoutineCompletionResultSubject.send(false)
             }
         }
+         */
     }
 }
